--- a/pyiron_workflow/create.py
+++ b/pyiron_workflow/create.py
@@ -6,7 +6,6 @@ from __future__ import annotations
 
 from abc import ABC
 from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
-from functools import lru_cache
 
 from executorlib import Executor as ExecutorlibExecutor
 from pyiron_snippets.dotdict import DotDict
@@ -44,35 +43,30 @@ class Creator(metaclass=Singleton):
         self.function_node = function_node
 
     @property
-    @lru_cache(maxsize=1)
     def standard(self):
         from pyiron_workflow.nodes import standard
 
         return standard
 
     @property
-    @lru_cache(maxsize=1)
     def for_node(self):
         from pyiron_workflow.nodes.for_loop import for_node
 
         return for_node
 
     @property
-    @lru_cache(maxsize=1)
     def macro_node(self):
         from pyiron_workflow.nodes.macro import macro_node
 
         return macro_node
 
     @property
-    @lru_cache(maxsize=1)
     def Workflow(self):
         from pyiron_workflow.workflow import Workflow
 
         return Workflow
 
     @property
-    @lru_cache(maxsize=1)
     def meta(self):
         from pyiron_workflow.nodes.transform import inputs_to_list, list_to_outputs
 
@@ -84,7 +78,6 @@ class Creator(metaclass=Singleton):
         )
 
     @property
-    @lru_cache(maxsize=1)
     def transformer(self):
         from pyiron_workflow.nodes.transform import (
             dataclass_node,
@@ -117,14 +110,12 @@ class Wrappers(metaclass=Singleton):
     as_function_node = staticmethod(as_function_node)
 
     @property
-    @lru_cache(maxsize=1)
     def as_macro_node(self):
         from pyiron_workflow.nodes.macro import as_macro_node
 
         return as_macro_node
 
     @property
-    @lru_cache(maxsize=1)
     def as_dataclass_node(self):
         from pyiron_workflow.nodes.transform import as_dataclass_node