--- a/pyiron_workflow/create.py
+++ b/pyiron_workflow/create.py
@@ -30,7 +30,6 @@ class Creator(metaclass=Singleton):
     """
 
     def __init__(self):
-
         # Standard lib
         self.ProcessPoolExecutor = ProcessPoolExecutor
         self.ThreadPoolExecutor = ThreadPoolExecutor