--- a/tests/integration/test_readme.py
+++ b/tests/integration/test_readme.py
@@ -16,5 +16,5 @@ class TestTriggerFromIDE(unittest.TestCase):
         pass
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()