--- a/tests/integration/test_workflow.py
+++ b/tests/integration/test_workflow.py
@@ -102,20 +102,18 @@ class TestWorkflow(unittest.TestCase):
         )
 
     def test_for_loop(self):
-
         base = 42
         to_add = list(range(5))
         bulk_loop = Workflow.create.for_node(
             demo_nodes.OptionallyAdd,
             iter_on="y",
             x=base,  # Broadcast
-            y=to_add  # Scattered
+            y=to_add,  # Scattered
         )
         out = bulk_loop()
 
         for output, expectation in zip(
-            out.df["sum"].values.tolist(),
-            [base + v for v in to_add], strict=False
+            out.df["sum"].values.tolist(), [base + v for v in to_add], strict=False
         ):
             self.assertAlmostEqual(
                 output,
@@ -151,7 +149,7 @@ class TestWorkflow(unittest.TestCase):
             Workflow.create.ProcessPoolExecutor,
             Workflow.create.ThreadPoolExecutor,
             Workflow.create.CloudpickleProcessPoolExecutor,
-            Workflow.create.ExecutorlibExecutor
+            Workflow.create.ExecutorlibExecutor,
         ]
 
         wf = Workflow("executed")
@@ -168,18 +166,20 @@ class TestWorkflow(unittest.TestCase):
             self.assertDictEqual(reference_output, reloaded.outputs.to_value_dict())
 
         for exe_cls in executors:
-            with self.subTest(
-                f"{exe_cls.__module__}.{exe_cls.__qualname__} entire workflow"
-            ), exe_cls() as exe:
+            with (
+                self.subTest(
+                    f"{exe_cls.__module__}.{exe_cls.__qualname__} entire workflow"
+                ),
+                exe_cls() as exe,
+            ):
                 wf.executor = exe
                 self.assertDictEqual(
-                    reference_output,
-                    wf().result().outputs.to_value_dict()
+                    reference_output, wf().result().outputs.to_value_dict()
                 )
                 self.assertFalse(
                     wf.running,
                     msg="The workflow should stop. For thread pool this required a "
-                        "little sleep"
+                        "little sleep",
                 )
             wf.executor = None
 
@@ -193,7 +193,7 @@ class TestWorkflow(unittest.TestCase):
                     any(n.running for n in wf),
                     msg=f"All children should be done running -- for thread pools this "
                         f"requires a very short sleep -- got "
-                        f"{[(n.label, n.running) for n in wf]}"
+                        f"{[(n.label, n.running) for n in wf]}",
                 )
             for child in wf:
                 child.executor = None
@@ -219,7 +219,7 @@ class TestWorkflow(unittest.TestCase):
             second_out,
             msg="Even thought the _input_ hasn't changed, we expect to avoid the first "
                 "(cached) result by virtue of resetting the cache when the body of "
-                "the composite graph has changed"
+                "the composite graph has changed",
         )
 
         t0 = time.perf_counter()
@@ -228,13 +228,13 @@ class TestWorkflow(unittest.TestCase):
         self.assertEqual(
             third_out,
             second_out,
-            msg="This time there is no change and we expect the cached result"
+            msg="This time there is no change and we expect the cached result",
         )
         self.assertLess(
             dt,
             0.1 * wf.c.inputs.t.value,
             msg="And because it used the cache we expect it much faster than the sleep "
-                "time"
+                "time",
         )
 
     def test_failure(self):
@@ -259,7 +259,10 @@ class TestWorkflow(unittest.TestCase):
         wf.starting_nodes = [wf.a]
         wf.automate_execution = False
 
-        with self.subTest("Check completion"), Workflow.create.ProcessPoolExecutor() as exe:
+        with (
+            self.subTest("Check completion"),
+            Workflow.create.ProcessPoolExecutor() as exe,
+        ):
             wf.c_fails.executor = exe
             wf(raise_run_exceptions=False)
 
@@ -270,7 +273,7 @@ class TestWorkflow(unittest.TestCase):
                 (wf.d_if_success.outputs.user_input.value, NOT_DATA),  # Never ran
                 (
                     wf.d_if_failure.outputs.user_input.value,
-                    wf.d_if_failure.inputs.user_input.value
+                    wf.d_if_failure.inputs.user_input.value,
                 ),
                 (wf.e_fails.outputs.add.value, NOT_DATA),
             ]:
@@ -297,12 +300,12 @@ class TestWorkflow(unittest.TestCase):
                     self.assertIn(
                         wf.c_fails.run.full_label,
                         str(e),
-                        msg="Failed node should be identified"
+                        msg="Failed node should be identified",
                     )
                     self.assertIn(
                         wf.e_fails.run.full_label,
                         str(e),
-                        msg="Indeed, _both_ failed nodes should be identified"
+                        msg="Indeed, _both_ failed nodes should be identified",
                     )
 
                 with self.subTest("Check recovery file"):
@@ -311,7 +314,7 @@ class TestWorkflow(unittest.TestCase):
                             filename=wf.as_path().joinpath("recovery")
                         ),
                         msg="Expect a recovery file to be written for the parent-most"
-                            "object when a child fails"
+                        "object when a child fails",
                     )
             finally:
                 wf.delete_storage()
@@ -322,9 +325,9 @@ class TestWorkflow(unittest.TestCase):
                         f"written a recovery file, so after removing that the whole "
                         f"node directory for the workflow should be cleaned up."
                         f"Instead, {wf.as_path()} exists and has content "
-                        f"{[f for f in wf.as_path().iterdir()] if wf.as_path().is_dir() else None}"
+                        f"{[f for f in wf.as_path().iterdir()] if wf.as_path().is_dir() else None}",
                 )
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()