--- a/tests/static/docs_submodule/good_function.py
+++ b/tests/static/docs_submodule/good_function.py
@@ -1,7 +1,7 @@
 def function():
     """
     Here is an A-OK docstring
-    
+
     >>> print(42)
     42
     """