--- a/tests/static/docs_submodule/mix.py
+++ b/tests/static/docs_submodule/mix.py
@@ -13,6 +13,7 @@ def bad():
     """
     return
 
+
 def error():
     """
     >>> 1/0