--- a/tests/static/nodes_subpackage/subsub_sibling/demo_nodes.py
+++ b/tests/static/nodes_subpackage/subsub_sibling/demo_nodes.py
@@ -2,7 +2,6 @@
 A demo node package for the purpose of testing.
 """
 
-
 from pyiron_workflow import Workflow