--- a/tests/unit/test_create.py
+++ b/tests/unit/test_create.py
@@ -8,5 +8,5 @@ class TestCreator(unittest.TestCase):
         Creator()
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()