I got stuck in authentification hell, so I only had time to just get started; more to come.

Timing test:

```python
from pyiron_workflow import Workflow

%%timeit
Workflow.create.transformer

```

With lru cache: `45.3 ns ± 0.584 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)`

Without: `46.6 ns ± 0.398 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)`

A completely inconsequential difference, so I got rid of the creator/wrapper caches. I'll check the remaining cache next.