@liamhuber 2024-12-21
@XzzX, with these amendments I'm happy with #520. Basically I accepted most of the formatting, but ultimately I turned it off and we'll lean on black for now. Otherwise the only anti-ruff thing I did was keep the top-level `Workflow` import un-sorted, and tell it to _always_ ignore unused imports in `__init__.py` files (since we use these to specify APIs).